requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2
djangorestframework==3.14.0
gunicorn==21.2.0
numpy==1.24.4
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3
shapely==2.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
